Renamed Fifo.h to VibratoBuffer.h
