+++ CMakeLists.txt
@@ -46,7 +46,7 @@ target_compile_features("${TARGET_NAME}" PRIVATE cxx_std_20)
 set(SourceFiles
     Source/Component/MidiMeter.h
     Source/VibratoDetector.h
-    Source/Utility/Fifo.h
+        Source/Utility/VibratoBuffer.h
     Source/PluginEditor.h
     Source/PluginEditor.cpp
     Source/PluginProcessor.h
