+++ Source/Utility/VibratoBuffer.h
