+++ Source/VibratoDetector.h
@@ -2,7 +2,7 @@
 
 #include "juce_audio_basics/juce_audio_basics.h"
 #include "juce_core/juce_core.h"
-#include "Utility/Fifo.h"
+#include "Utility/VibratoBuffer.h"
 
 class VibratoDetector {
 public:
