+++ Tests/CMakeLists.txt
@@ -20,7 +20,7 @@ set(UnitTests
 # A list of all the files that we need for our unit tests. Unless you write tests for every single class
 # used in the application, it is probably better to just include the files we actually need
 set(TestedFiles
-        ../Source/Utility/Fifo.h)
+        ../Source/Utility/VibratoBuffer.h)
 
 juce_add_console_app(MyPluginTests
         PRODUCT_NAME "MyPlugin Unit Tests")
